requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain==0.1.13
langchain-openai==0.1.1
chromadb==0.4.24
python-dotenv==1.0.1
python-multipart==0.0.9
sentence-transformers==2.5.1